Выделяет модули reset и send-data
